fix: Create Card Flow with selected boardId and Listid

+++ core/data/src/main/java/com/jetapps/jettaskboard/local/dao/DashboardDao.kt
@@ -32,5 +32,5 @@ interface DashboardDao {
     fun getAllLists(): List<ListEntity>
 
     @Query("SELECT * FROM listTable where boardId = :boardId")
-    fun getAllBoardRelatedLists(boardId: Int): List<ListEntity>
+    fun getAllBoardRelatedLists(boardId: Long): List<ListEntity>
 }
