+++ core/data/src/main/java/com/jetapps/jettaskboard/local/source/DatabaseSource.kt
@@ -13,7 +13,7 @@ interface DatabaseSource {
     suspend fun getBoard(boardId : Long) : Flow<BoardWithLists?>
     suspend fun getBoards() : Flow<List<BoardEntity>>
     suspend fun getAllLists() : List<ListEntity>
-    suspend fun getAllListsRelatedToBoard(boardId : Int) : List<ListEntity>
+    suspend fun getAllListsRelatedToBoard(boardId : Long) : List<ListEntity>
     suspend fun updateCard(cardEntity: CardEntity)
     suspend fun createCard(cardEntity: CardEntity)
     suspend fun deleteCard(cardEntity: CardEntity)
